requirements.txt
typing_extensions~=4.8.0
uvicorn~=0.23.0
websocket-client~=1.6.2
docker~=7.0.0
fastapi~=0.109.2
requests~=2.31.0
kubernetes~=29.0.0
pandas~=1.5.3
numpy~=1.26.0
plotly~=5.19.0
streamlit~=1.31.1
python-dotenv~=0.21.0
httpx~=0.27.0
starlette~=0.36.0
yaml~=0.2.5
pyyaml~=6.0.1
pydantic~=2.7.1
dxlib~=1.0.21+20.git.g202e295
setuptools~=68.0.0
websockets~=10.4